Stopped the LocationMethods from automatically processing the JSON data to a modifed Hashie::Mash. Processing of large amounts of JSON in pure Ruby is too slow, even for a plain Hashie::Mash. The Mash class can now be used more selectively by client code.
Modified the Mash class to retain the original key format for keys related to forecast data as these match the values in the wx params.

Now explicitly using the Oj gem to parse JSON returned by the API into a Ruby hash.

Added a fugly script to collect JSON test data fixtures from the API and removed the VCR gem. Further simplified testing by removing MiniTest::Metadata.

This version completes testing of the location-specific data API endpoints and the conversion of the JSON to Mash.